# spring-server/src/main/java/com/example/spring_server/chat/dto/ConversationDTO.java
package com.example.spring_server.chat.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.LocalDateTime;

/**
 * ConversationDTO
 * 
 * Data Transfer Object for Conversation.
 * Used in REST API & service layer to avoid exposing internal entity structure.
 */
@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class ConversationDTO {
    private Long id;
    private Long customerId;
    private String channel;
    private String status;
    private LocalDateTime createdAt;
    private LocalDateTime updatedAt;
    private LocalDateTime lastMessageTime;
    private Long messageCount;
}